Moved remaining operations into their own files.
