+++ db/tables/ddl/split.py
@@ -1,0 +1,118 @@
+from sqlalchemy import Column, func, select, ForeignKey, literal, exists
+
+from db import columns, constants
+from db.tables.ddl.create import create_mathesar_table
+from db.tables.utils import reflect_table
+
+
+def _split_column_list(columns_, extracted_column_names):
+    extracted_columns = [
+        col for col in columns_ if col.name in extracted_column_names
+    ]
+    remainder_columns = [
+        col for col in columns_ if col.name not in extracted_column_names
+    ]
+    return extracted_columns, remainder_columns
+
+
+def _create_split_tables(extracted_table_name, extracted_columns, remainder_table_name, remainder_columns, schema, engine):
+    extracted_table = create_mathesar_table(
+        extracted_table_name,
+        schema,
+        extracted_columns,
+        engine,
+    )
+    remainder_fk_column = Column(
+        f"{extracted_table.name}_{constants.ID}",
+        columns.ID_TYPE,
+        ForeignKey(f"{extracted_table.name}.{constants.ID}"),
+        nullable=False,
+    )
+    remainder_table = create_mathesar_table(
+        remainder_table_name,
+        schema,
+        [remainder_fk_column] + remainder_columns,
+        engine,
+        metadata=extracted_table.metadata
+    )
+    return extracted_table, remainder_table, remainder_fk_column.name
+
+
+def _create_split_insert_stmt(old_table, extracted_table, extracted_columns, remainder_table, remainder_columns, remainder_fk_name):
+    SPLIT_ID = f"{constants.MATHESAR_PREFIX}_split_column_alias"
+    extracted_column_names = [col.name for col in extracted_columns]
+    remainder_column_names = [col.name for col in remainder_columns]
+    split_cte = select(
+        [
+            old_table,
+            func.dense_rank().over(order_by=extracted_columns).label(SPLIT_ID)
+        ]
+    ).cte()
+    cte_extraction_columns = (
+        [split_cte.columns[SPLIT_ID]]
+        + [split_cte.columns[n] for n in extracted_column_names]
+    )
+    cte_remainder_columns = (
+        [split_cte.columns[SPLIT_ID]]
+        + [split_cte.columns[n] for n in remainder_column_names]
+    )
+    extract_sel = select(
+        cte_extraction_columns,
+        distinct=True
+    )
+    extract_ins_cte = (
+        extracted_table
+        .insert()
+        .from_select([constants.ID] + extracted_column_names, extract_sel)
+        .returning(literal(1))
+        .cte()
+    )
+    remainder_sel = select(
+        cte_remainder_columns,
+        distinct=True
+    ).where(exists(extract_ins_cte.select()))
+
+    split_ins = (
+        remainder_table
+        .insert()
+        .from_select(
+            [remainder_fk_name] + remainder_column_names,
+            remainder_sel
+        )
+    )
+    return split_ins
+
+
+def extract_columns_from_table(old_table_name, extracted_column_names, extracted_table_name, remainder_table_name, schema, engine, drop_original_table=False):
+    old_table = reflect_table(old_table_name, schema, engine)
+    old_columns = (
+        columns.MathesarColumn.from_column(col) for col in old_table.columns
+    )
+    old_non_default_columns = [
+        col for col in old_columns if not col.is_default
+    ]
+    extracted_columns, remainder_columns = _split_column_list(
+        old_non_default_columns, extracted_column_names,
+    )
+    with engine.begin() as conn:
+        extracted_table, remainder_table, remainder_fk = _create_split_tables(
+            extracted_table_name,
+            extracted_columns,
+            remainder_table_name,
+            remainder_columns,
+            schema,
+            engine,
+        )
+        split_ins = _create_split_insert_stmt(
+            old_table,
+            extracted_table,
+            extracted_columns,
+            remainder_table,
+            remainder_columns,
+            remainder_fk,
+        )
+        conn.execute(split_ins)
+    if drop_original_table:
+        old_table.drop()
+
+    return extracted_table, remainder_table, remainder_fk
