+++ db/tests/records/conftest.py
@@ -3,15 +3,14 @@ import pytest
 from sqlalchemy import MetaData, Table
 
 
-ROSTER = "Roster"
 FILTER_SORT = "filter_sort"
 
 
 @pytest.fixture
-def roster_table_obj(engine_with_roster):
+def roster_table_obj(engine_with_roster, roster_table_name):
     engine, schema = engine_with_roster
     metadata = MetaData(bind=engine)
-    roster = Table(ROSTER, metadata, schema=schema, autoload_with=engine)
+    roster = Table(roster_table_name, metadata, schema=schema, autoload_with=engine)
     return roster, engine
