+++ db/tests/test_columns.py
@@ -11,7 +11,6 @@ from sqlalchemy import (
 from sqlalchemy.exc import IntegrityError
 from db import columns, constants, constraints
 from db.tables import ddl as table_ddl
-from db.tables import operations as table_operations
 from db.tables import utils as table_utils
 from db.types import email, alteration
 from db.types.base import get_db_type_name
@@ -287,7 +286,7 @@ def test_rename_column_foreign_keys(engine_with_schema):
     table_name = "table_to_split"
     columns_list = [Column("Filler 1", Integer), Column("Filler 2", Integer)]
     table_ddl.create_mathesar_table(table_name, schema, columns_list, engine)
-    extracted, remainder, fk_name = table_operations.extract_columns_from_table(
+    extracted, remainder, fk_name = table_ddl.extract_columns_from_table(
         table_name, ["Filler 1"], "Extracted", "Remainder", schema, engine
     )
     new_fk_name = "new_" + fk_name
